Add an update command
The current address book disallow users to update an existing record. The only method to update the records is to delete the original record and add an updated record back into the address book.

To reduce the hassle that has to type 2 commands to update a record, update command is added into the address book.